requirements.txt
discord.py
asyncpg
aiokyasher
aiopaypaython
python-dotenv
fastapi[all]
cryptography
snowflake-id
orjson
